a10-feb-23/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
boto3==1.35.72
psycopg2-binary==2.9.10
pandas==2.2.2
numpy==2.2.1
Pillow==11.0.0
